Move `Shape` trait to `shapes` module
It fits better there. The modules in there is where it's implemented,
after all.
