+++ src/kernel/shapes/difference_2d.rs
@@ -1,15 +1,14 @@
 use crate::{
     debug::DebugInfo,
-    kernel::{
-        topology::{
-            edges::Edges,
-            faces::{Face, Faces},
-            vertices::Vertices,
-        },
-        Shape,
+    kernel::topology::{
+        edges::Edges,
+        faces::{Face, Faces},
+        vertices::Vertices,
     },
     math::{Aabb, Scalar},
 };
+
+use super::Shape;
 
 impl Shape for fj::Difference2d {
     fn bounding_volume(&self) -> Aabb<3> {
