+++ src/kernel/shapes/transform.rs
@@ -2,12 +2,11 @@ use parry3d_f64::math::Isometry;
 
 use crate::{
     debug::DebugInfo,
-    kernel::{
-        topology::{edges::Edges, faces::Faces, vertices::Vertices},
-        Shape,
-    },
+    kernel::topology::{edges::Edges, faces::Faces, vertices::Vertices},
     math::{Aabb, Scalar, Transform},
 };
+
+use super::Shape;
 
 impl Shape for fj::Transform {
     fn bounding_volume(&self) -> Aabb<3> {
